+++ frontend/src/pages/UserAuth/Register.tsx
@@ -1,31 +1,46 @@
 import React, { useState } from "react";
-import { signIn } from "../services/firebaseAuthService";
+import { register } from "../../services/firebaseAuthService";
 
-const Login: React.FC = () => {
+const Register: React.FC = () => {
+  const [name, setName] = useState("");
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
+  const [confirmPassword, setConfirmPassword] = useState("");
   const [error, setError] = useState("");
-  const [showPassword, setShowPassword] = useState(false);
+  const [showPassword, setShowPassword] = useState(false); // State to toggle password visibility
 
   const handleSubmit = async (event: React.FormEvent) => {
     event.preventDefault();
     setError("");
 
+    if (password !== confirmPassword) {
+      setError("Passwords do not match");
+      return;
+    }
+
     try {
-      const token = await signIn(email, password);
-      if (!token) throw new Error("No token received");
-      localStorage.setItem("jwt", token);
+      await register(email, password);
     } catch (err: any) {
-      setError(err.message);
+      setError(err.message || "An unexpected error occurred");
     }
   };
 
   return (
     <div className="flex items-center justify-center h-screen bg-orange-500">
       <div className="w-full max-w-md p-8 space-y-6 bg-white rounded-lg shadow-lg">
-        <h2 className="text-3xl font-bold text-center text-black">Login</h2>
+        <h2 className="text-3xl font-bold text-center text-black">Register</h2>
         {error && <p className="text-red-500 text-center">{error}</p>}
         <form className="space-y-4" onSubmit={handleSubmit}>
+          <div>
+            <label className="block mb-1 text-sm font-medium text-black">Name</label>
+            <input
+              type="text"
+              className="w-full px-4 py-2 border rounded-lg"
+              value={name}
+              onChange={(e) => setName(e.target.value)}
+              required
+            />
+          </div>
           <div>
             <label className="block mb-1 text-sm font-medium text-black">Email</label>
             <input
@@ -53,28 +68,30 @@ const Login: React.FC = () => {
               {showPassword ? "Hide" : "Show"}
             </button>
           </div>
-          <div className="flex items-center justify-between">
-            <label className="flex items-center">
-              <input type="checkbox" className="mr-2" />
-              <span className="text-sm text-black">Remember me</span>
+          <div>
+            <label className="block mb-1 text-sm font-medium text-black">
+              Confirm Password
             </label>
-            <a href="/reset-password" className="text-sm text-orange-600 hover:underline">
-              Forgot Password?
-            </a>
-
+            <input
+              type="password"
+              className="w-full px-4 py-2 border rounded-lg"
+              value={confirmPassword}
+              onChange={(e) => setConfirmPassword(e.target.value)}
+              required
+            />
           </div>
           <button
             type="submit"
             className="w-full px-4 py-2 text-white bg-orange-600 rounded-lg hover:bg-orange-700 transition ease-in-out duration-200"
-          >
-            Log In
-          </button>
+            >
+         Register
+        </button>
 
         </form>
         <p className="text-center text-sm text-black">
-          Don’t have an account?{" "}
-          <a href="/register" className="text-orange-600 hover:underline">
-            Register
+          Already have an account?{" "}
+          <a href="/login" className="text-orange-600 hover:underline">
+            Log In
           </a>
         </p>
       </div>
@@ -82,4 +99,4 @@ const Login: React.FC = () => {
   );
 };
 
-export default Login;
+export default Register;
