+++ frontend/src/routes/appRoutes.tsx
@@ -2,10 +2,10 @@ import React from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import Home from "../pages/Home";
 import Catalog from "../pages/Catalog/Catalog"
-import Login from "../pages/Login";
-import Register from "../pages/Register";
-import ResetPassword from "../pages/resetPassword";
-import ResetPasswordForm from "../pages/resetPasswordForm";
+import Login from "../pages/UserAuth/Login";
+import Register from "../pages/UserAuth/Register";
+import ResetPassword from "../pages/UserAuth/resetPassword";
+import ResetPasswordForm from "../pages/UserAuth/resetPasswordForm";
 
 const AppRoutes: React.FC = () => (
   <Router>
